Update reachability flags when the app returns to the foreground since we don't receive notifications in the background.
Because the run loop doesn't run in the background, we never receive reachability change notifications when in the background. This could result in the object's cached flags being out of date when you return to the foreground. Fix this by adding a listener to the UIApplicationWillEnterForeground notification and updating flags then.

Additionally, only fire KVO and NSNotifications if reachability has actually changed. If flags are the exact same, no notification/KVO is fired.